Fixing scala style
