+++ functional-tests/src/main/scala/com/vertica/spark/functests/endtoend/ComplexTypeTests.scala
@@ -4,7 +4,7 @@ import com.vertica.spark.config.{FileStoreConfig, JDBCConfig}
 import com.vertica.spark.datasource.jdbc.VerticaJdbcLayer
 import com.vertica.spark.functests.TestUtils
 import com.vertica.spark.util.error.{ComplexTypeReadNotSupported, ConnectorException, InternalMapNotSupported}
-import com.vertica.spark.util.schema.{MetadataKey, SchemaTools}
+import com.vertica.spark.util.schema.{MetadataKey, ComplexTypeSchemaSupport}
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{DataFrame, Row, SaveMode}
 
@@ -196,7 +196,7 @@ class ComplexTypeTests(readOpts: Map[String, String], writeOpts: Map[String, Str
       val columnRs = stmt.executeQuery(s"select data_type_id from columns where table_name='$tableName' and column_name='$colName'")
       assert(columnRs.next)
       val verticaId = columnRs.getLong("data_type_id")
-      assert(verticaId > SchemaTools.VERTICA_SET_BASE_ID & verticaId < SchemaTools.VERTICA_SET_MAX_ID)
+      assert(verticaId > ComplexTypeSchemaSupport.VERTICA_SET_BASE_ID & verticaId < ComplexTypeSchemaSupport.VERTICA_SET_MAX_ID)
     } catch {
       case err : Exception => fail(err)
     }
